Added logic in order to process embedded $parent and $this refs
New Logic tries to compare a $this or $parent references not found in tree by increments of the ref string. This is now similar to a common $root call and allows for the $this or $parent string to be embedded into another string and still be evaluated. The current implementation may be a little bit crude but does the job. Further performance tuning may be required, maybe during _extract_tree refactoring.